move connection handling to net lib
