+++ Cargo.toml
@@ -4,4 +4,5 @@ members = [
     "rust-ms-login",
     "packet",
     "crypt",
+    "net"
 ]
