+++ net/src/lib.rs
@@ -1,0 +1,10 @@
+
+pub mod session;
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        assert_eq!(2 + 2, 4);
+    }
+}
